Get current stats impl.

+++ modules/explorer-api/src/main/scala/org/ergoplatform/explorer/http/api/v0/models/StatsSummary.scala
@@ -1,13 +1,15 @@
 package org.ergoplatform.explorer.http.api.v0.models
 
+import scala.math.BigDecimal
+
 final case class StatsSummary(
   blocksCount: Long,
   blocksAvgTime: Long,
   totalCoins: Long,
   totalTransactionsCount: Long,
   totalFee: Long,
-  totalOutput: Long,
-  estimatedOutput: BigInt,
+  totalOutput: BigDecimal,
+  estimatedOutput: BigDecimal,
   totalMinerRevenue: Long,
   percentEarnedTransactionsFees: Double,
   percentTransactionVolume: Double,
